add: Option to suppress checksum validation of incoming data from ve.direct data-stream
Option to suppress checksum validation of incoming data from ve.direct data-stream. This allows to read incoming data from battery chargers of the Phoenix Smart IP43 family and leads to a clean, uninterrupted data stream of voltages and current. This mofification was tested for about half a year. This feature was actually already implemented but a corresponding checkbox was missing in the user interface. Additionally now logging checksum errors to syslog is suppressed as well.