Report V/I as solar rather than battery if MPPT is connected
VE.Direct uses the same protocol for battery monitors and MPPT
controllers (and presumably chargers too), but the voltage and current
it reports using the 'V' and 'I' type refer to quite different things
depending on whether a battery monitor or a solar charger is connected.

In the former case, the voltage and current are the battery voltage and
current at the shunt, whereas in the latter, they are the voltage and
current coming out of the charger. In the case of the voltage, the
difference is unlikely to be noticeable, assuming they are connected to
the same battery bank. But the battery current and the charge current
are really not the same thing: they differ by the amount of current that
is consumed in the network, along with current coming from other
chargers.

This means that reporting the current as 'electrical.batteries.*.current'
is never correct when the source is a MPPT controller, and its output
may conflict with the output of a BMV when it reports information on the
same battery bank.

The correct thing to do is to report 'electrical.solar.*.current' in
this case, which describes the current sent into the circuit by the
MPPT.

In order to reduce the risk of disruption of existing installations,
infer the device type from the 'bmv' and 'solar' settings: if 'solar' is
set and 'bmv' is cleared (which never happens by default, and is highly
unlikely to be set in this manner unless the source in question is a
MMPT), report the V and I values as 'electrical.solar.<name>.voltage'
and 'electrical.solar.<name>.current', where the name is taken from the
'solar' setting in the plugin config.