Implement PUT handler for relay state when connected to serial
Implement support for manipulating the state of the BMV-7xx builtin
relay via PUT messages, making it available as a generic switch for
other plugins.

Note that this requires configuring the relay as remotely controlled via
the BMV control panel.

Implemented for serial connections only - it should be straight-forward
to do the same for TCP but it is doubtful whether this is desirable.